+++ src/types/v1/CSN-Interop-Effective.ts
@@ -136,28 +136,88 @@ export type PersonalDataIsPotentiallySensitive = boolean;
  */
 export type SemanticsCurrencyCode = true;
 /**
- * The property contains a unit of measure.
+ * The element contains a unit of measure.
  */
 export type SemanticsUnitOfMeasure = true;
+/**
+ * The element states the day of month.
+ */
 export type SemanticsCalendarDayOfMonth = true;
+/**
+ * The element states the day of year.
+ */
 export type SemanticsCalendarDayOfYear = true;
+/**
+ * The element states the calendar week.
+ */
 export type SemanticsCalendarWeek = true;
+/**
+ * The element states the calendar month.
+ */
 export type SemanticsCalendarMonth = true;
+/**
+ * The element states the calendar quarter.
+ */
 export type SemanticsCalendarQuarter = true;
+/**
+ * The element states the calendar halfyear.
+ */
 export type SemanticsCalendarHalfyear = true;
+/**
+ * The element states the calendar year.
+ */
 export type SemanticsCalendarYear = true;
+/**
+ * The element states the calendar year week.
+ */
 export type SemanticsCalendarYearWeek = true;
+/**
+ * The element states the calendar year month.
+ */
 export type SemanticsCalendarYearMonth = true;
+/**
+ * The element states the calendar year quarter.
+ */
 export type SemanticsCalendarYearQuarter = true;
+/**
+ * The element states the calendar year halfyear.
+ */
 export type SemanticsCalendarYearHalfyear = true;
+/**
+ * The element states the calendar year variant.
+ */
 export type SemanticsFiscalYearVariant = true;
+/**
+ * The element states the fiscal period.
+ */
 export type SemanticsFiscalPeriod = true;
+/**
+ * The element states the fiscal year.
+ */
 export type SemanticsFiscalYear = true;
+/**
+ * The element states the fiscal year period.
+ */
 export type SemanticsFiscalYearPeriod = true;
+/**
+ * The element states the fiscal quarter.
+ */
 export type SemanticsFiscalQuarter = true;
+/**
+ * The element states the fiscal year quarter.
+ */
 export type SemanticsFiscalYearQuarter = true;
+/**
+ * The element states the fiscal week.
+ */
 export type SemanticsFiscalWeek = true;
+/**
+ * The element states the fiscal year week.
+ */
 export type SemanticsFiscalYearWeek = true;
+/**
+ * The element states the fiscal day of year.
+ */
 export type SemanticsFiscalDayOfYear = true;
 /**
  * The property contains a language code.
@@ -173,16 +233,10 @@ export type SemanticsTime = true;
 export type SemanticsText = true;
 /**
  * Tags a field containing a GUID in RAW16 or CHAR32 format.
+ *
+ * It is RECOMMENDED to explicitly use the `csd.UUID` type.
  */
 export type SemanticsUuid = true;
-/**
- * The property contains a date which is the lower boundary of a (validity) interval (for time dependent master data).
- */
-export type SemanticsBusinessDateFrom = true;
-/**
- * The property contains a date which is the upper boundary of a (validity) interval (for time dependent master data).
- */
-export type SemanticsBusinessDateTo = true;
 /**
  * Total number of digits that are present after the decimal point in a number.
  * The scale can hold from zero up to the total numeric precision
@@ -196,6 +250,14 @@ export type DecimalScaleNumber = number;
 export type DecimalScaleType = DecimalScaleType1 & DecimalScaleType2;
 export type DecimalScaleType1 = "floating";
 export type DecimalScaleType2 = string;
+/**
+ * The property contains a date which is the lower boundary of a (validity) interval (for time dependent master data).
+ */
+export type SemanticsBusinessDateFrom = true;
+/**
+ * The property contains a date which is the upper boundary of a (validity) interval (for time dependent master data).
+ */
+export type SemanticsBusinessDateTo = true;
 /**
  * The operator "="
  */
@@ -412,10 +474,12 @@ export interface Definitions {
   [k: string]: DefinitionEntry;
 }
 /**
- * A context corresponds to a CAP CSN Namespace. It is used to create a bounded context that bundles elements of different kinds together.
- * The context will become part of the entity / element name.
+ * A context corresponds to a CAP CSN Namespace.
+ * It is used to create a bounded context that bundles elements of different kinds together.
  *
- * The context is also useful to avoid local or global name collisions.
+ * To assign Services, Entities or Custom Types to a context, their name MUST be prefixed with the context name, separated by a `.`.
+ *
+ * The context can also be useful to avoid local or global name collisions.
  *
  * See [Primer: Context](../primer.md#context-definitions).
  */
@@ -706,32 +770,6 @@ export interface BooleanType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -943,8 +981,6 @@ export interface StringType {
   "@Semantics.time"?: SemanticsTime;
   "@Semantics.text"?: SemanticsText;
   "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1062,32 +1098,6 @@ export interface LargeStringType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1155,6 +1165,7 @@ export interface IntegerType {
   "@PersonalData.fieldSemantics"?: PersonalDataFieldSemantics & PersonalDataFieldSemantics1;
   "@PersonalData.isPotentiallyPersonal"?: PersonalDataIsPotentiallyPersonal;
   "@PersonalData.isPotentiallySensitive"?: PersonalDataIsPotentiallySensitive;
+  "@Semantics.valueRange"?: Semantics;
   "@Semantics.currencyCode"?: SemanticsCurrencyCode;
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
@@ -1179,12 +1190,6 @@ export interface IntegerType {
   "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
   "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
   "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1211,6 +1216,31 @@ export interface DefaultValueInteger {
    * Default Value for Integer Type.
    */
   val: number | null;
+}
+/**
+ * Informs consumers about the value range of a CDS element, allowing to specify minimal and/or maximal values, and indicate whether these are exclusive or inclusive.
+ * This value range can be used for client-side validation of user input.
+ */
+export interface Semantics {
+  /**
+   * Specifies the minimum value.
+   */
+  minimum?: string;
+  /**
+   * Specifies if the lower boundary should be excluded.
+   * Not specifying this annotation means that the lower boundary is included.
+   */
+  exclusiveMinimum?: boolean;
+  /**
+   * Specifies the maximum value.
+   */
+  maximum?: string;
+  /**
+   * Specifies if the upper boundary should be excluded.
+   * Not specifying this annotation means that the upper boundary is included.
+   */
+  exclusiveMaximum?: boolean;
+  [k: string]: unknown | undefined;
 }
 /**
  * Signed integer with 64 bit.
@@ -1258,36 +1288,11 @@ export interface Integer64Type {
   "@PersonalData.fieldSemantics"?: PersonalDataFieldSemantics & PersonalDataFieldSemantics1;
   "@PersonalData.isPotentiallyPersonal"?: PersonalDataIsPotentiallyPersonal;
   "@PersonalData.isPotentiallySensitive"?: PersonalDataIsPotentiallySensitive;
+  "@Semantics.valueRange"?: Semantics;
   "@Semantics.currencyCode"?: SemanticsCurrencyCode;
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1361,36 +1366,11 @@ export interface DecimalType {
   "@PersonalData.fieldSemantics"?: PersonalDataFieldSemantics & PersonalDataFieldSemantics1;
   "@PersonalData.isPotentiallyPersonal"?: PersonalDataIsPotentiallyPersonal;
   "@PersonalData.isPotentiallySensitive"?: PersonalDataIsPotentiallySensitive;
+  "@Semantics.valueRange"?: Semantics;
   "@Semantics.currencyCode"?: SemanticsCurrencyCode;
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1459,36 +1439,11 @@ export interface DoubleType {
   "@PersonalData.fieldSemantics"?: PersonalDataFieldSemantics & PersonalDataFieldSemantics1;
   "@PersonalData.isPotentiallyPersonal"?: PersonalDataIsPotentiallyPersonal;
   "@PersonalData.isPotentiallySensitive"?: PersonalDataIsPotentiallySensitive;
+  "@Semantics.valueRange"?: Semantics;
   "@Semantics.currencyCode"?: SemanticsCurrencyCode;
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1560,30 +1515,6 @@ export interface DateType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
   "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
   "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
@@ -1657,32 +1588,6 @@ export interface TimeType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1754,32 +1659,6 @@ export interface DateTimeType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1851,32 +1730,6 @@ export interface TimestampType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -1947,32 +1800,6 @@ export interface UUIDType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -2057,32 +1884,6 @@ export interface AssociationType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -2215,32 +2016,6 @@ export interface CompositionType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -2348,32 +2123,6 @@ export interface CustomType {
   "@Semantics.amount.currencyCode"?: ElementReference;
   "@Semantics.unitOfMeasure"?: SemanticsUnitOfMeasure;
   "@Semantics.quantity.unitOfMeasure"?: ElementReference;
-  "@Semantics.calendar.dayOfMonth"?: SemanticsCalendarDayOfMonth;
-  "@Semantics.calendar.dayOfYear"?: SemanticsCalendarDayOfYear;
-  "@Semantics.calendar.week"?: SemanticsCalendarWeek;
-  "@Semantics.calendar.month"?: SemanticsCalendarMonth;
-  "@Semantics.calendar.quarter"?: SemanticsCalendarQuarter;
-  "@Semantics.calendar.halfyear"?: SemanticsCalendarHalfyear;
-  "@Semantics.calendar.year"?: SemanticsCalendarYear;
-  "@Semantics.calendar.yearWeek"?: SemanticsCalendarYearWeek;
-  "@Semantics.calendar.yearMonth"?: SemanticsCalendarYearMonth;
-  "@Semantics.calendar.yearQuarter"?: SemanticsCalendarYearQuarter;
-  "@Semantics.calendar.yearHalfyear"?: SemanticsCalendarYearHalfyear;
-  "@Semantics.fiscal.yearVariant"?: SemanticsFiscalYearVariant;
-  "@Semantics.fiscal.period"?: SemanticsFiscalPeriod;
-  "@Semantics.fiscal.year"?: SemanticsFiscalYear;
-  "@Semantics.fiscal.yearPeriod"?: SemanticsFiscalYearPeriod;
-  "@Semantics.fiscal.quarter"?: SemanticsFiscalQuarter;
-  "@Semantics.fiscal.yearQuarter"?: SemanticsFiscalYearQuarter;
-  "@Semantics.fiscal.week"?: SemanticsFiscalWeek;
-  "@Semantics.fiscal.yearWeek"?: SemanticsFiscalYearWeek;
-  "@Semantics.fiscal.dayOfYear"?: SemanticsFiscalDayOfYear;
-  "@Semantics.language"?: SemanticsLanguage;
-  "@Semantics.time"?: SemanticsTime;
-  "@Semantics.text"?: SemanticsText;
-  "@Semantics.uuid"?: SemanticsUuid;
-  "@Semantics.businessDate.from"?: SemanticsBusinessDateFrom;
-  "@Semantics.businessDate.to"?: SemanticsBusinessDateTo;
   /**
    * Annotations or private properties MAY be added.
    *
@@ -2602,7 +2351,7 @@ export interface SupportedCapabilitiesEnumValue {
 /**
  * A CDS Service indicates that [Entities](#entity-definition) are exposed to outside consumers via an API.
  *
- * To assign Entities to a Service, the entity name MUST be prefixed with the service name, separated by a `.`.
+ * To assign Entities or Custom Types to a Service, their name MUST be prefixed with the service name, separated by a `.`.
  * Please be aware that this will also change the `target` values of `cds.Association` and `cds.Composition`.
  * For an example, see [airline-service.json](./examples/airline.md).
  *